requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.13.0

# Financial Data
yfinance>=0.2.28
ccxt>=4.0.0
pandas-datareader>=0.10.0

# Econometrics & Time Series
arch>=6.2.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dashboard
streamlit>=1.28.0
streamlit-aggrid>=0.3.4

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
